+++ src/main/java/plugily/projects/murdermystery/handlers/language/LanguageMigrator.java
@@ -34,7 +34,7 @@ import java.io.File;
 public class LanguageMigrator {
 
   public static final int CONFIG_FILE_VERSION = 18;
-  public static final int LANGUAGE_FILE_VERSION = 3;
+  public static final int LANGUAGE_FILE_VERSION = 4;
   private final Main plugin;
 
   public LanguageMigrator(Main plugin) {
@@ -214,10 +214,17 @@ public class LanguageMigrator {
             "  Inv-Title: \"Arena selector\"\r\n" +
             "  Item:\r\n" +
             "    Lore:\r\n" +
-            "      - \"&4Murder Mystery\"\r\n" +
-            "      - \"&7%state%\"\r\n" +
-            "      - \"&c%mapname%\"\r\n" +
-            "      - \"&5[%playersize%/%maxplayers%]\"");
+            "      - \"&4Murder Mystery &f- &e%mapname%\"\r\n" +
+            "      - \" \"\r\n" +
+            "      - \" \"\r\n" +
+            "      - \"  &fOnline: %playersize%/%maxplayers%\"\r\n" +
+            "      - \"  &fState: %state%\"\r\n" +
+            "      - \" \"\r\n" +
+            "      - \" \"\r\n" +
+            "      - \"&eClick to join this arena\"\r\n");
+          break;
+        case 3:
+          MigratorUtils.insertAfterLine(file, "In-Game:", "  Game-Death-Format: \"&7[&4☠&7] &r\"");
           break;
         default:
           break;