+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.6.2-b2 Beta
+* Fixed goldrush & goldban prayer did not reset on game end
+* Death prefix can now be changed
+
 ### 1.6.2-b Beta
 * Fixed MYSQL stats can't update
 * (Developer Alert - Changed class names)
@@ -10,6 +14,8 @@
 * Added ArenaSelector
 * Reworked version checker
 * Changed message delay of "Waiting for players" from 15 to 45 seconds 
+* Fixed no permission message not found
+* Optimized player corpse heads on paper
 
 ### 1.6.1 Release (01.07.2020 - 27.07.2020)
 * Added missing 1.14 & 1.15 & 1.16 Materials